requirements.txt
telethon==1.35.1
python-dotenv==1.0.0
Pillow==10.1.0
opencv-python==4.8.1.78